Add guards to the role for multi-node clouds
This change adds a missing endif block in the cgroup template and sets a
default for the nova virt type should it be undefined. This change also
adds guards around the haproxy and scripts task files so that both local
scripts to this role and upstream osa scripts can be used at the same
time. The sudo template was updated to ensure success by making sure
required variables are defined.

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>